101. Symmetric Tree
"Given a binary tree, check whether it is a mirror of itself (ie, symmetric around its center)."
-----
Optimal solution that uses BFS with TWO queues, where items are inserted in reverse order.
Thus in each node, the children should be in same order, and if they are not then binary tree is not mirrored.
Also, if the queue lengths are different, they can't be mirrors.