Biweekly Contest #32 - 1541. Minimum Insertions to Balance a Parentheses String
"Given a parentheses string s containing only the characters '(' and ')'. A parentheses string is balanced if:

Any left parenthesis '(' must have a corresponding two consecutive right parenthesis '))'.
Left parenthesis '(' must go before the corresponding two consecutive right parenthesis '))'.
For example, "())", "())(())))" and "(())())))" are balanced, ")()", "()))" and "(()))" are not balanced.

You can insert the characters '(' and ')' at any position of the string to balance it if needed.

Return the minimum number of insertions needed to make s balanced."

-----
Solutions uses concept of stack and peeking next character. '(' are always pushed to stack and popped
appropriately to fill the conditions of preceding ')' characters. This solution wasn't submitted on time
for the contest; had a wrong assumption at one point.