Longest Palindrome
"Given a string which consists of lowercase or uppercase letters, find the length of the longest palindromes that can be built with those letters.

This is case sensitive, for example "Aa" is not considered a palindrome here."
---
Solution:
Create histogram and consider even count of each character.
One character can have odd count (potential middle character in odd-length
string).