Sum of Nodes with Even-Valued Grandparent
"Given a binary tree, return the sum of values of nodes with even-valued grandparent.  (A grandparent of a node is the parent of its parent, if it exists.)

If there are no nodes with an even-valued grandparent, return 0."
-----
Solution:
Based on BFS (queue)
Utilizing Swift's optional chaining and nil-coalescing operator to shorten the code.