HARD: Minimum Number of Days to Eat N Oranges
"There are n oranges in the kitchen and you decided to eat some of these oranges every day as follows:

Eat one orange.
If the number of remaining oranges (n) is divisible by 2 then you can eat  n/2 oranges.
If the number of remaining oranges (n) is divisible by 3 then you can eat  2*(n/3) oranges.
You can only choose one of the actions per day.

Return the minimum number of days to eat n oranges.
"
-----
Solution:
Use Dynamic Programming/Recursion to find the mimimum combination. However,
This solution is good only for limited (<50,0000) oranges. The requirement is to
able to handle 2B oranges. This was my submission in Leet Contest #202, but it
failed in test case of large number of oranges.