Numbers With Same Consecutive Differences
"Return all non-negative integers of length N such that the absolute difference between every two consecutive digits is K.

Note that every number in the answer must not have leading zeros except for the number 0 itself. For example, 01 has one leading zero and is invalid, but 0 is valid.

You may return the answer in any order."

-----
Solution:
You may at first think this is a simple for-loop algorithm, but this actually requires Dynamic Programming
approach (recursive or iterative).
Following is a recursive solution.