HARD: Stream of Characters
"Implement the StreamChecker class as follows:

StreamChecker(words): Constructor, init the data structure with the given words.
query(letter): returns true if and only if for some k >= 1, the last k characters queried (in order from oldest to newest, including this letter just queried) spell one of the words in the given list.
"
----
Solution:
Naiive approach:
keep a list of words that match the stream of characters.

While this solution works, it doesn't pass large test sets due to time out.
The problem author assumes that Trie -structure is used to solve this problem.