HARD: Stream of Characters
"Implement the StreamChecker class as follows:

StreamChecker(words): Constructor, init the data structure with the given words.
query(letter): returns true if and only if for some k >= 1, the last k characters queried (in order from oldest to newest, including this letter just queried) spell one of the words in the given list.
"
----
Solution:
Use Trie.

While this solution works, it doesn't pass large test sets due to time out.
It turns out that for performance reasons, we need to enter all words in REVERSE order to our trie.
That's the trick. Otherwise the solution is way too slow for given test sets.