HARD: Stream of Characters
PASSING SOLUTION
"Implement the StreamChecker class as follows:

StreamChecker(words): Constructor, init the data structure with the given words.
query(letter): returns true if and only if for some k >= 1, the last k characters queried (in order from oldest to newest, including this letter just queried) spell one of the words in the given list.
"
----
Solution:
Use Trie, but reverse the stream and the words. This allows to exit the loop much faster.
As per discussions in LeetCode, this approach will also Timeout with certain test material
(thus reversing doesn't provide a solution). Reversing works when there's a long word that
differs from the stream by the last characters (we can bail out quickly). Not an ideal
LeetCode problem as there are room for interpretation and test cases are inadequate (vs. constraints).
Also, the hint provided in the LC problem guides most people to implement this in non-optimal
manner.