4Sum
"Given an array nums of n integers and an integer target, are there elements a, b, c, and d in nums such that a + b + c + d = target? Find all unique quadruplets in the array which gives the sum of target.

Note:

The solution set must not contain duplicate quadruplets."

------
Solution:
 Use two nested loops o(n^2) nested with two pointer traversal.
 Execution complexity ~o(n^3)
Don't mess with hash maps; gets way too complicated and is harder to understand.