Find Right Interval
"Given a set of intervals, for each of the interval i, check if there exists an interval j whose start point is bigger than or equal to the end point of the interval i, which can be called that j is on the "right" of i.

For any interval i, you need to store the minimum interval j's index, which means that the interval j has the minimum start point to build the "right" relationship for interval i. If the interval j doesn't exist, store -1 for the interval i. Finally, you need output the stored value of each interval as an array.

Note:

You may assume the interval's end point is always bigger than its start point.
You may assume none of these intervals have the same start point."

----
Solution:
Create a new struct that contains the start, end and index of each given interval. Then sort these structs based on
start. Then simply search for end until it is found (one could use Binary Search here to accelerate it, but I resorted
to not use it for simplicity.)