Implement Rand10() using Rand7()
"Given a function rand7 which generates a uniform random integer in the range 1 to 7, write a function rand10 which generates a uniform random integer in the range 1 to 10.

Do NOT use system's Math.random()."

-----
Solution:
Naiive approach would be to first attempt merely using two calls to Rand(7), sum up and multiple by 10 and divide by 14.
This is far from stochastic solution.

Second would be to again use two calls; multiple by 10 to get range 10-70; then add randomness by adding another rand7()*10/7, finally
divide by 7 to get range 1-10, which is far better, but still not stochastic enough.

We need proper randomness distribution. This can be achieved reliably only by using a mapping (and rejection of mapping attempts that
are outside of the stochastic map.). Otherwise can't achieve true randomness of the scaling results from 7 -> 10.
Keep in mind, we are dealing with INTEGERS here. If it were Floating point values, the scaling would work.

It is true that there's a simple one-liner that can be stochastic enough:
return (rand7()+rand7()+rand7()+rand7()+rand7()+rand7()+rand7()+rand7()+rand7()+rand7()) % 10 + 1
But this requires multiple calls to rand7.