Pancake Sort
"
Given an array of integers A, We need to sort the array performing a series of pancake flips.

In one pancake flip we do the following steps:

Choose an integer k where 0 <= k < A.length.
Reverse the sub-array A[0...k].
For example, if A = [3,2,1,4] and we performed a pancake flip choosing k = 2, we reverse the sub-array [3,2,1], so A = [1,2,3,4] after the pancake flip at k = 2.

Return an array of the k-values of the pancake flips that should be performed in order to sort A. Any valid answer that sorts the array within 10 * A.length flips will be judged as correct."

-----
Solution:
First find out where is the current largest number (index).
Flip the numbers so current largest number becomes first; then flip again
so current largest number goes tothe  back of the currently sorted numbers. This way we are
sorting from the end of the array until we come to the first item. At each step we will check if we are
in sorted order already.