HARD: Largest Component Size by Common Factor
(LeetCode rating Hard)
"Given a non-empty array of unique positive integers A, consider the following graph:

There are A.length nodes, labelled A[0] to A[A.length - 1];
There is an edge between A[i] and A[j] if and only if A[i] and A[j] share a common factor greater than 1.
Return the size of the largest connected component in the graph."

-----
Solution:
First calculate primenumbers for first 320 numbers (root of 100,000 is ~320; 100,000 is max number size).
Then calculate which numbers have a common factor between them. Create one-way or two-way graph.

Then recursively traverse through all the nodes to find the longest path.

This solution works but is O(N^N) complex, so does not work with large number sets. Alternative approach is required
(theory says to use Topolgical sort).