Delete Node in a BST
"Given a root node reference of a BST and a key, delete the node with the given key in the BST. Return the root node reference (possibly updated) of the BST.

Basically, the deletion can be divided into two stages:

Search for a node to remove.
If the node is found, delete the node.
Note: Time complexity should be O(height of tree)."

-----
Solution:
This requires the use of theory.
Few cases: Node to delete: is a leaf, has one child. has two children.
First two cases are simple, but the case with two children requires knowledge. Theory says that we have to find the next minimum value to the
RIGHT of the node with the key. Once found, exchange values and delete the minimum node (could be a leaf or has one child).