Contains Duplicate II
"Given an array of integers and an integer k, find out whether there are two distinct indices i and j in the array such that nums[i] = nums[j] and the absolute difference between i and j is at most k."

----
Solution:
uses pre-sorting as suggested in "Contains Duplicate III". However, simpler solution is to store
index of each number while traversing through the array. If there is an index stored for a given
number already in the hash map, check if that index is within K limits. If not, replace that value
with current values index and move on to the next number.