Partition Labels
"A string S of lowercase English letters is given. We want to partition this string into as many parts as possible so that each letter appears in at most one part, and return a list of integers representing the size of these parts."

----
Solution:
Use list of map of characters.