Maximum Depth of Binary Tree
"Given a binary tree, find its maximum depth.

The maximum depth is the number of nodes along the longest path from the root node down to the farthest leaf node.

Note: A leaf is a node with no children."
-----
Solution:
Use BFS to determine levels based on number of children in queue at any level.