Intersection of Two Linked Lists
"Write a program to find the node at which the intersection of two singly linked lists begins."

-----
Solution:
Using lengths of both lists, but reversed list and calculating the position of the
first common node. Working, relatively fast, but overly complex. The proper algorithm
from LeetCode is in commented out code section in the end.