Image Overlap
"Two images A and B are given, represented as binary, square matrices of the same size.  (A binary matrix has only 0s and 1s as values.)

We translate one image however we choose (sliding it left, right, up, or down any number of units), and place it on top of the other image.  After, the overlap of this translation is the number of positions that have a 1 in both images.

(Note also that a translation does not include any kind of rotation.)

What is the largest possible overlap?"

---------
Solution:
Important assupmtion: IMAGES DO NOT HAVE TO MATCH; we just want to know how much they overlap by 1's.

"Brute-force" by sliding one image over the other left to right, top-to-bottom; This covers over overlapping scenarios.
This could be improved by focusing on the overlapping window only.