Word Pattern
"Given a pattern and a string str, find if str follows the same pattern.

Here follow means a full match, such that there is a bijection between a letter in pattern and a non-empty word in str."

---------
Solution:
Use two Hash maps (dictionaries); one for pattern char, one for word.