Merge K Sorted Lists
LeetCode HARD
"You are given an array of k linked-lists lists, each linked-list is sorted in ascending order.

Merge all the linked-lists into one sorted linked-list and return it."

---------
Solution:
Initial solution TLEd, so I came up with this rather simple solution. We'll strip all nodes from their lists and put them in an array which we sort. Then we recompile the returned linked list from these nodes. This is considered "Brute force", but still beats 80% of Swift submissions. Should Swift have Priority Queues as part of standard libraries, this would be rather elegant to solve with those... But we don't have them.