Jump Game
"Given an array of non-negative integers, you are initially positioned at the first index of the array.

Each element in the array represents your maximum jump length at that position.

Determine if you are able to reach the last index."

---------
Solution:
Instead of using recursion or memo, we can just simply keep track of the last good position
that allows us to reach the end index.