Combinatino Sum
"Given an array of distinct integers candidates and a target integer target, return a list of all unique combinations of candidates where the chosen numbers sum to target. You may return the combinations in any order.

The same number may be chosen from candidates an unlimited number of times. Two combinations are unique if the frequency of at least one of the chosen numbers is different."

----------
Solution:
 Backtracking using DFS.
 Key here is to avoid navigating to indexes in combinations which
 we have investigated. That's why helper keeps track of the index. This
 enables us to neatly avoid duplicates as well. Also, values are kept outside of
 the helper as Array copy operations can be costly. (for each iteration we restore
 the values candidate in the helper.