Sort Colors
"Given an array nums with n objects colored red, white, or blue, sort them in-place so that objects of the same color are adjacent, with the colors in the order red, white, and blue.

Here, we will use the integers 0, 1, and 2 to represent the color red, white, and blue respectively.

Follow up:

Could you solve this problem without using the library's sort function?
Could you come up with a one-pass algorithm using only O(1) constant space?"

----------
Solution:
Calculating the number of 0,1,2s and then writing over the nums with the number of 0,1,2s accordingly.
Two-passes (O(n)) complexity.