Remove Covered Intervals
"Given a list of intervals, remove all intervals that are covered by another interval in the list.

Interval [a,b) is covered by interval [c,d) if and only if c <= a and b <= d.

After doing so, return the number of remaining intervals."

---------
Solution:
While sorting would provide more optimal approach for larger data sets (o (nlogn)), in this case, the brute force'd o(n^2) solution
is actually faster to execute. Beats 100% of Swift submission by a wide marginn.