Repeated DNA Sequences
"All DNA is composed of a series of nucleotides abbreviated as 'A', 'C', 'G', and 'T', for example: "ACGAATTCCG". When studying DNA, it is sometimes useful to identify repeated sequences within the DNA.

Write a function to find all the 10-letter-long sequences (substrings) that occur more than once in a DNA molecule."

---------
Solution:

All else fails but use of hash map. Since there are 100,000 characters or so to go trough, we can't create an array of characters either (too slow).
But iterate over the given string. Create 10-char long hash map entries for each starting position; This is o(n) time.