Minimum Domino Rotations for Equal Row
"In a row of dominoes, A[i] and B[i] represent the top and bottom halves of the ith domino.  (A domino is a tile with two numbers from 1 to 6 - one on each half of the tile.)

We may rotate the ith domino, so that A[i] and B[i] swap values.

Return the minimum number of rotations so that all the values in A are the same, or all the values in B are the same.

If it cannot be done, return -1."

---------
Solution:
At first calculated histogram, but that is not necessary. We can do this in one-pass and keep track
of both bottom and top rows at the same time. Simple and easy.